Fixes lightbox that did not show correctly on mobile NDLA-248
